+++ runtime/playground/public/examples/expand_index_select.py
@@ -3,7 +3,7 @@ import torch
 
 x = torch.tensor([[1.0, 2.0], [3.0, 4.0]])
 e = x.expand(3, 2, 2)
-indices = torch.tensor([0, 1], dtype=torch.int64)
+indices = torch.tensor([0, 1], dtype=torch.int32)
 s = torch.index_select(x, 0, indices)
 out = {
   "input": x.tolist(),
